Use double as default real number precision.

+++ src/real.h
@@ -1,13 +1,6 @@
-/*
- * real.h
- *
- *  Created on: May 1, 2012
- *      Author: geir
- */
-
 #ifndef REAL_H_
 #define REAL_H_
 
-typedef float real;
+typedef double real;
 
 #endif /* REAL_H_ */
